// src/abi/types/EnvoysSale.ts
export type SaleInfoStruct = {
  token: PromiseOrValue<string>;
  soft: PromiseOrValue<BigNumberish>;
  hard: PromiseOrValue<BigNumberish>;
  buyToken: PromiseOrValue<string>;
  price: PromiseOrValue<BigNumberish>;
  start: PromiseOrValue<BigNumberish>;
  end: PromiseOrValue<BigNumberish>;
};